gcloud/appengine/pyproject.toml
[tool.poetry]
name = "appengine"
version = "0.1.0"
description = ""
authors = ["i544c <leo@i544c.me>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
